fix checkDuplicate trigger tests
