+++ test/Data.java
@@ -38,8 +38,14 @@ public class Data {
         public static final Set<String> RMDupperTest__resolveDuplicate_reverse_merged_expectedReadNames = Stream.of(
         "M_NS500559:30:HNM25BGXX:2:12204:22389:14660"
         ).collect(Collectors.toSet());
-        public static final Set<String> RMDupperTest__queueOrOutput_nextAlignmentStartBeyondFirstEndTriggersCheckForDuplication_expectedReadNames = Stream.of(
+        public static final Set<String> RMDupperTest__queueOrOutput_nextAlignmentStartBeyondFirstEndTriggersCheckForDuplication_pre_expectedReadNames = Stream.of(
+        "M_NS500559:30:HNM25BGXX:1:12108:16299:13403",
         "M_NS500559:30:HNM25BGXX:3:12601:6521:5315",
+        "M_NS500559:30:HNM25BGXX:2:12204:22389:14659"
+        ).collect(Collectors.toSet());
+        public static final Set<String> RMDupperTest__queueOrOutput_nextAlignmentStartBeyondFirstEndTriggersCheckForDuplication_post_expectedReadNames = Stream.of(
+        "M_NS500559:30:HNM25BGXX:3:12601:6521:5315",
+        "M_NS500559:30:HNM25BGXX:2:12204:22389:14659",
         "M_NS500559:30:HNM25BGXX:2:12204:22389:14660"
         ).collect(Collectors.toSet());
 }
